Remove MCP and MCF, replace with MouseClickAction
